dev: move the bip324 files to the external submodule
