+++ src/external/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod bip324;
 pub(crate) mod chacha20_poly1305;
